Better scaling, still needs work
Implemented a maximum size for the orbs which fixed the infinite growth with flashlight selection issue. Orb scaling works perfectly with bubble cursor, but has issues with flashlight. In addition, when you switch to flashlight with resized orbs, they return back to their original scale (1.0, 1.0, 1.0). Not sure what causes this, probably hard coded values in the flashlight script.