Added scaling population
Population changes based on the scale of the orb. When the scale is within a certain range, the population will be set to a specific value.
Additionally, I think I got the VR stuff to work. I'm going to test it more later. I want to see if scaling is added.