Fixed Scaling
Still some minor issues:
1. When you are holding an orb and switch to a different controller, the object you are selecting is dropped.
2. If you select an object with flashlight, and then select the same object with another flashlight controller, the original position is set to  directly in front of the player

There are probably other issues I have not found, but as far as I am able to tell, scaling is working as intended.

Also, the connections between orbs needs to be updated based on the scale, because they are currently shooting straight through the orbs, leading to a rather ugly web.