Test scene is working
The test scene is working properly. I created copies of bubble cursor and flashlight that have been modified to suit the GUI better. They should also work better with other functionality if added because the controllers aren't responsible for everything. I'm going to add this stuff to JMain to have everything together and modify the scripts accordingly tomorrow. I'm also going to update some things like flashlight because there are better ways to implement it that are easier to fix if something goes wrong.