Added orb labels
Orbs have a label that appears when highlighting them. I'm going to try and add a billboard effect to these. I should be able to just add it with the godot billboard property, but I also want to add this to the GUI which works differently. I'm going to test this all tomorrow once it's working.